Backing up.  I have been working on the z sorter, and I think most code is there now, although it hasn't been tested.  Single sided facets are still not supported, and they actually make sense in a z sorter (as opposed to in the z buffer).
git-svn-id: https://svn.isy.liu.se/rt/tidefelt/source/drool/trunk@330 c1f300ed-6a01-0410-bd5d-8e39f192a3d9

+++ source/corepoint.cc
@@ -29,7 +29,7 @@ Lang::Core_duration::call( Kernel::EvalState * evalState, Kernel::Arguments & ar
       typedef const Lang::ElementaryPath2D ArgType;
       RefCountPtr< ArgType > arg = Helpers::elementaryPathTry2D( args.getValue( 0 ) );
       Kernel::ContRef cont = evalState->cont_;
-      cont->takeValue( Kernel::ValueRef( new Lang::Float( arg->duration( ) ) ),
+      cont->takeValue( Kernel::ValueRef( new Lang::Integer( arg->duration( ) ) ),
 		       evalState );
       return;
     }
@@ -44,7 +44,7 @@ Lang::Core_duration::call( Kernel::EvalState * evalState, Kernel::Arguments & ar
       typedef const Lang::ElementaryPath3D ArgType;
       RefCountPtr< ArgType > arg = Helpers::elementaryPathTry3D( args.getValue( 0 ) );
       Kernel::ContRef cont = evalState->cont_;
-      cont->takeValue( Kernel::ValueRef( new Lang::Float( arg->duration( ) ) ),
+      cont->takeValue( Kernel::ValueRef( new Lang::Integer( arg->duration( ) ) ),
 		       evalState );
       return;
     }
